+++ app/permissions/decidim/decidim_awesome/admin/permissions.rb
@@ -12,20 +12,16 @@ module Decidim
           return permission_action if user.read_attribute("admin").blank?
           return permission_action unless permission_action.action == :edit_config
 
-          if permission_action.subject == :admin_accountability && DecidimAwesome.admin_accountability.respond_to?(:include?)
-            if global?
-              toggle_allow(DecidimAwesome.admin_accountability.include?(:admin_roles))
-            else
-              toggle_allow(DecidimAwesome.admin_accountability.include?(:participatory_space_roles))
-            end
-          elsif permission_action.subject == :private_data && config_enabled?(:proposal_private_custom_fields)
-            if private_data.present?
-              allow! if private_data.destroyable?
+          apply_private_data_permissions!
+          if config_enabled?(*permission_action.subject)
+            case permission_action.subject
+            when :admins_available_authorizations
+              apply_admin_authorizations_permissions!
+            when :admin_accountability
+              apply_admin_accountability_permissions!
             else
               allow!
             end
-          else
-            toggle_allow(config_enabled?(*permission_action.subject))
           end
 
           permission_action
@@ -33,12 +29,47 @@ module Decidim
 
         private
 
+        def apply_admin_authorizations_permissions!
+          allow! if awesome_admin_authorizations.present? && handler.in?(awesome_admin_authorizations)
+        end
+
+        def apply_admin_accountability_permissions!
+          if DecidimAwesome.admin_accountability.respond_to?(:include?)
+            if global?
+              allow! if DecidimAwesome.admin_accountability.include?(:admin_roles)
+            elsif DecidimAwesome.admin_accountability.include?(:participatory_space_roles)
+              allow!
+            end
+          end
+        end
+
+        def apply_private_data_permissions!
+          return unless permission_action.subject == :private_data && config_enabled?(:proposal_private_custom_fields)
+
+          if private_data.present?
+            allow! if private_data.destroyable?
+          else
+            allow!
+          end
+        end
+
         def global?
           context.fetch(:global, nil)
         end
 
         def private_data
           context.fetch(:private_data, nil)
+        end
+
+        def handler
+          context.fetch(:handler, "")
+        end
+
+        def awesome_admin_authorizations
+          @awesome_admin_authorizations ||= begin
+            handlers = AwesomeConfig.find_by(var: :admins_available_authorizations, organization: user.organization)&.value || []
+            user.organization.available_authorizations.intersection(handlers)
+          end
         end
       end
     end